Extract async source resolver
